requirements.txt
qiskit
quimb
cuquantum-python
networkx
scikit-learn
torch
numpy
Pillow
